Update dev scripts Remove markdown-it plugins
