+++ ts/packages/agents/markdown/package.json
@@ -20,13 +20,8 @@
     "postbuild": "copyfiles -u 1 \"src/**/config.json\" dist",
     "build:frontend": "vite build",
     "clean": "rimraf --glob dist *.tsbuildinfo *.done.build.log",
-    "dev": "node scripts/dev-viewer.js",
-    "dev:backend": "npm run tsc && node scripts/dev-viewer.js",
-    "dev:backend-no-file": "npm run tsc && node scripts/dev-viewer-no-file.js",
-    "dev:frontend": "vite --host",
-    "dev:full": "concurrently \"npm run dev:backend\" \"npm run dev:frontend\"",
-    "dev:full-no-file": "concurrently \"npm run dev:backend-no-file\" \"npm run dev:frontend\"",
-    "dev:hmr": "node scripts/dev-viewer-hmr.js",
+    "dev": "npm run tsc && node scripts/dev-viewer.js --hmr",
+    "dev:frontend": "vite  --watch",
     "prettier": "prettier --check . --ignore-path ../../../.prettierignore",
     "prettier:fix": "prettier --write . --ignore-path ../../../.prettierignore",
     "tsc": "tsc -b"
